requirements.txt
Django==4.2
django-environ==0.10
psycopg2==2.9.6
psycopg2-binary==2.9
whitenoise==6.4
django-nested-admin==4.0.2
python-monkey-business==1.0.0
six==1.16.0
setuptools==67.8.0
